examples/recognition/requirements.txt
requests==2.23.0
tqdm==4.46.0
numpy==1.16.6
pandas==0.24.2
Pillow==2.2.1
pathlib==1.0.1